Restore Shoot Cert Service observability
972a04e removed the gep19Monitoring variable from the helm chart, leading
to the extension only deploying the pre-GEP-19-style observability
ConfigMap, which contains a scrape config, an alerting rule, and the
Plutono dashboard.
It should also have removed its usage so that the extension helm chart
includes the new ConfigMap, containing only the Plutono dashboard, as well
as the ServiceMonitor CRD (to instruct Prometheus to scrape the
cert-controller-manager) and a PrometheusRule CRD containing the alerting
rule.